enhance hvigor plugin
