+++ crates/tauri-cli/src/mobile/mod.rs
@@ -105,7 +105,7 @@ impl Target {
       Self::Android => "android",
       #[cfg(target_os = "macos")]
       Self::Ios => "ios",
-      Self::OpenHarmony => "open-harmony",
+      Self::OpenHarmony => "ohos",
     }
   }
